Hard code instance types
Using variables to select instance types for EC2 and RDS instances in
CloudFormation was cumbersome and verbose for little overall gain.

Since we have upload the template from our local machines when creating new
stacks anyway, we can change the values or EC2 and RDS instance types, Image
IDs, and SnapshotIdentifier locally, and use them to build test stacks.

If and when test stacks prove to be the successful recipe, we can then commit
the hardcoded changes.

Less fancy, but also less error prone.